requirements.txt
#python version that i use is 3.9.6
ffmpeg-python 
opencv-python